Remove PyTorch dependency for TensorFlow and MindSpore workloads.
