+++ .github/workflows/mindspore_tests.yml
@@ -30,7 +30,6 @@ jobs:
           sudo apt-get update
           sudo apt install -y python3-pip
           conda activate plato
-          conda install pytorch torchvision cpuonly -c pytorch -y
           conda install mindspore-cpu=1.5.0 -c mindspore -c conda-forge -y
           pip3 install numpy python-socketio aiohttp requests boto3 pyyaml pyyaml-include datasets
