+++ .github/workflows/tensorflow_tests.yml
@@ -23,7 +23,6 @@ jobs:
                   sudo apt-get update
                   sudo apt install -y python3-pip
                   pip3 install -r requirements.txt --upgrade
-                  pip3 install torch==1.10.0+cpu torchvision==0.11.1+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html
                   pip3 install tensorflow tensorflow-datasets
 
             - name: Training workloads
