+++ examples/split_learning/split_learning_algorithm.py
@@ -14,6 +14,19 @@ from plato.utils import optimizers
 from plato.algorithms import fedavg
 
 
+class FeatureDataset(torch.utils.data.Dataset):
+    """Used to prepare a feature dataset for a DataLoader in PyTorch."""
+    def __init__(self, dataset):
+        self.dataset = dataset
+
+    def __len__(self):
+        return len(self.dataset)
+
+    def __getitem__(self, item):
+        image, label = self.dataset[item]
+        return image, label
+
+
 class Algorithm(fedavg.Algorithm):
     """The PyTorch-based split learning algorithm, used by both the client and the
     server.
@@ -111,4 +124,4 @@ class Algorithm(fedavg.Algorithm):
             self.client_id, toc - tic))
 
     def train(self, trainset, sampler, cut_layer=None):
-        self.trainer.train(trainset.torch_dataset(), sampler, cut_layer)
+        self.trainer.train(FeatureDataset(trainset), sampler, cut_layer)
