+++ plato/processors/registry.py
@@ -10,23 +10,28 @@ from collections import OrderedDict
 from typing import Tuple
 
 from plato.config import Config
-from plato.processors import (base, mistnet_inbound_features,
-                              mistnet_outbound_features,
-                              mistnet_randomized_response, mistnet_gaussian,
-                              mistnet_laplace, mistnet_unbatch,
-                              mistnet_quantize, mistnet_dequantize, pipeline)
+from plato.processors import pipeline
 
-registered_processors = OrderedDict([
-    ('base', base.Processor),
-    ('mistnet_randomized_response', mistnet_randomized_response.Processor),
-    ('mistnet_unbatch', mistnet_unbatch.Processor),
-    ('mistnet_outbound_features', mistnet_outbound_features.Processor),
-    ('mistnet_inbound_features', mistnet_inbound_features.Processor),
-    ('mistnet_gaussian', mistnet_gaussian.Processor),
-    ('mistnet_laplace', mistnet_laplace.Processor),
-    ('mistnet_quantize', mistnet_quantize.Processor),
-    ('mistnet_dequantize', mistnet_dequantize.Processor),
-])
+if not (hasattr(Config().trainer, 'use_tensorflow')
+        or hasattr(Config().trainer, 'use_mindspore')):
+    from plato.processors import (base, mistnet_inbound_features,
+                                  mistnet_outbound_features,
+                                  mistnet_randomized_response,
+                                  mistnet_gaussian, mistnet_laplace,
+                                  mistnet_unbatch, mistnet_quantize,
+                                  mistnet_dequantize)
+
+    registered_processors = OrderedDict([
+        ('base', base.Processor),
+        ('mistnet_randomized_response', mistnet_randomized_response.Processor),
+        ('mistnet_unbatch', mistnet_unbatch.Processor),
+        ('mistnet_outbound_features', mistnet_outbound_features.Processor),
+        ('mistnet_inbound_features', mistnet_inbound_features.Processor),
+        ('mistnet_gaussian', mistnet_gaussian.Processor),
+        ('mistnet_laplace', mistnet_laplace.Processor),
+        ('mistnet_quantize', mistnet_quantize.Processor),
+        ('mistnet_dequantize', mistnet_dequantize.Processor),
+    ])
 
 
 def get(user: str, *args,
