+++ plato/servers/registry.py
@@ -7,13 +7,13 @@ Having a registry of all available classes is convenient for retrieving an insta
 import logging
 from collections import OrderedDict
 
+from plato.config import Config
+
 from plato.servers import (
     fedavg,
     fedavg_cs,
     mistnet,
 )
-
-from plato.config import Config
 
 registered_servers = OrderedDict([
     ('fedavg', fedavg.Server),
@@ -31,7 +31,8 @@ def get(model=None, algorithm=None, trainer=None):
 
     if server_type in registered_servers:
         logging.info("Server: %s", server_type)
-        registered_server = registered_servers[server_type](model=model, algorithm=algorithm, trainer=trainer)
+        registered_server = registered_servers[server_type](
+            model=model, algorithm=algorithm, trainer=trainer)
     else:
         raise ValueError('No such server: {}'.format(server_type))
